Make lineplot support >1 games on same date
Games sharing the same date are offset by a certain number of hours.
The logic for the offset is as follows:

Let N be the number of games on a given date. For each game
n = 0, 1, ..., N - 1, the date is offset by n * 24 / (N + 1) hours.

This lets all games' datetimes have the same spacing, while having them
close to the original date (rather than the next day).

Hopefully this explanation isn't too obtuse.