Move the page-up button to top of screen
I also was able to have it appear not when the games list is first
visible, but a little while after when most of the games list occupies
the screen, which is nice. This was accomplished by messing with the
rootMargin option, which I somewhat understand. I *think* this solution
is robust.