Adjust number of line plot y-ticks
This is more of a band-aid solution than a proper one: the number of
y-ticks I pass to D3 seems to be more of a "suggestion" than an exact
prescription. At some point if I find myself needing to apply more of
these types of fixes, I should figure out a way to set the "suggestion"
so that I consistently get the results I want, regardless of the data.