Remove circular dependency from colour gen script
The data generation script can't process new data which includes new
players until it gets colours for the new players. The colour generation
script pulls its data from the output of the data generation script,
which will not include the new players (since it can't run until it has
colours).

So with the current system it would be impossible to add new players,
at least doing so using the scripts I've written. The solution that's
included in this commit makes it so the colour generation script pulls
data directly from the raw data (not the data output of the data
generation script). Now when you add new players, you run the script to
get colours, then you run the data generation script.