Fix bug when having more than one game per date
This is easier to describe with an example: Say we have player X and
player Y. On a given day, X plays games 1 and 2. Y only plays game 2. On
the lineplot, X's data point would be offset, as it should, because it's
for the second game of the day. However, Y's data point would not be
offset, which is the bug.

So instead of looking globally at what games were played on the same
dates, I instead did it on a per-player basis. That's fixed now.