Add fallback options when no game counts are met
The people section and the boxplot require a minimum number of games per
player for at least one player. If this condition isn't met, the people
section throws an error (bad), while the boxplot simply doesn't render
anything (less bad, but still bad). Now, if the condition isn't met, all
players will be considered, even if it makes basically zero sense for
the boxplot (in terms of it being statistically/interpretively useful).